third_party/blink/web_tests/platform/generic/fast/events/button-mouse-active-expected.txt

button
PASS successfullyParsed is true

TEST COMPLETE

PASS getComputedStyle(button).backgroundColor is "rgb(0, 255, 0)"
PASS getComputedStyle(button).backgroundColor is "rgb(0, 0, 255)"
